+++ src/epp/domain/update_domain.cc
@@ -168,7 +168,7 @@ unsigned long long update_domain(
                 parameter_value_range_errors.add_extended_error(
                         EppExtendedError::of_vector_parameter(
                                 Param::domain_ext_val_date,
-                                boost::numeric_cast<unsigned short>(0),
+                                0,
                                 Reason::valexpdate_not_valid));
             }
             else
@@ -192,7 +192,7 @@ unsigned long long update_domain(
                     parameter_value_range_errors.add_extended_error(
                             EppExtendedError::of_vector_parameter(
                                     Param::domain_ext_val_date,
-                                    boost::numeric_cast<unsigned short>(0),
+                                    0,
                                     Reason::valexpdate_not_valid));
                 }
             }
@@ -202,7 +202,7 @@ unsigned long long update_domain(
     }
     else   // not enum
     {
-        if(_update_domain_data.enum_validation)
+        if (_update_domain_data.enum_validation)
         {
             parameter_value_range_errors.add_extended_error(
                     EppExtendedError::of_vector_parameter(
